refactor: move legacy model out of the way for the new model
