+++ app/metrics/models/__init__.py
@@ -1,0 +1,3 @@
+from .common import *
+from .questions import *
+from .legacy import *
