+++ app/metrics/models/legacy.py
@@ -1,167 +1,11 @@
 from django.db import models
-from django.contrib.postgres.fields import ArrayField
-from django.urls import reverse
-from django import forms
+from .common import ChoiceArrayField, string_choices, EditTracking
 from django.core.exceptions import ValidationError
-import re
-import requests
-from django.contrib.auth.models import User 
+from django.contrib.auth.models import User
 
 
-def string_choices(choices):
-    return [
-        (choice, choice)
-        for choice in choices
-    ]
-
-
-class ChoiceArrayField(ArrayField):
-    def formfield(self, **kwargs):
-        defaults = {
-            "form_class": forms.MultipleChoiceField,
-            "choices": self.base_field.choices,
-        }
-        defaults.update(kwargs)
-        return super(ArrayField, self).formfield(**defaults)
-
-class Event(models.Model):
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    created = models.DateTimeField(auto_now_add=True)
-    modified = models.DateTimeField(auto_now=True)
-    code = models.CharField(max_length=128, unique=True, null=True, blank=True)
-    title = models.CharField(max_length=1024)
-    node = models.ManyToManyField("Node")
-    node_main = models.ForeignKey(
-        "Node", on_delete=models.CASCADE, related_name='node_main')
-    date_start = models.DateField()
-    date_end = models.DateField()
-    duration = models.DecimalField(verbose_name="Duration (days)", max_digits=6, decimal_places=2)
-    type = models.TextField(
-        choices=string_choices([
-            "Training - face to face",
-            "Training - e-learning",
-            "Training - blended",
-            "Knowledge Exchange Workshop",
-            "Hackathon",
-        ])
-    )
-    organising_institution = models.ManyToManyField("OrganisingInstitution")
-    location_city = models.CharField(max_length=128)
-    location_country = models.CharField(max_length=128)
-    funding = ChoiceArrayField(base_field=models.TextField(
-        choices=string_choices([
-            "ELIXIR Converge",
-            "EOSC Life",
-            "EXCELERATE",
-            "ELIXIR Implementation Study",
-            "ELIXIR Community / Use case",
-            "ELIXIR Node",
-            "ELIXIR Hub",
-            "ELIXIR Platform",
-            "Non-ELIXIR / Non-EXCELERATE Funds",
-        ])
-    ))
-
-
-    target_audience = ChoiceArrayField(base_field=models.TextField(
-        choices=string_choices([
-            "Academia/ Research Institution",
-            "Industry",
-            "Non-Profit Organisation",
-            "Healthcare",
-        ])
-    ))
-
-    additional_platforms = ChoiceArrayField(base_field=models.TextField(
-        choices=string_choices([
-            "Compute",
-            "Data",
-            "Interoperability",
-            "Tools",
-            "NA",
-        ])
-    ))
-
-    communities = ChoiceArrayField(base_field=models.TextField(
-        choices=string_choices([
-            "Human Data",
-            "Marine Metagenomics",
-            "Rare Diseases",
-            "Plant Sciences",
-            "Proteomics",
-            "Metabolomics",
-            "Galaxy",
-            "NA",
-        ])
-    ))
-
-    number_participants = models.PositiveIntegerField()
-    number_trainers = models.PositiveIntegerField()
-    url = models.URLField(max_length=4096)
-    status = models.TextField(
-        choices=string_choices([
-            "Complete",
-            "Incomplete",
-        ])
-    )
-    locked = models.BooleanField(default=False)
-
-    def __str__(self):
-        return f"{self.title} ({self.id}) ({self.code})"
-    
-
-    def get_absolute_url(self):
-        return reverse("event-edit", kwargs={"pk": self.id})
-    
-
-    @property
-    def location(self):
-        return (
-            self.location_city,
-            self.location_country
-        )
-    
-    @property
-    def date_period(self):
-        return (
-            self.date_start,
-            self.date_end
-        )
-
-    @property
-    def stats(self):
-        return [
-            (name, model.objects.filter(event=self).count())
-            for name, model in [
-                ("Quality metrics", Quality),
-                ("Impact metrics", Impact),
-                ("Demographic metrics", Demographic)
-            ]
-        ]
-
-    @property
-    def metrics_status(self):
-        count = sum([1 if v > 0 else 0 for _n, v in self.stats])
-        return {
-            0: "None",
-            1: "Partial",
-            2: "Partial",
-            3: "Full"
-        }[count]
-    
-    @property
-    def is_locked(self):
-        return (
-            self.code is not None
-            or self.locked
-        )
-
-
-class Demographic(models.Model):
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    created = models.DateTimeField(auto_now_add=True)
-    modified = models.DateTimeField(auto_now=True)
-    event = models.ForeignKey("Event", on_delete=models.CASCADE)
+class Demographic(EditTracking):
+    event = models.ForeignKey("Event", on_delete=models.CASCADE, related_name="demographic")
     employment_country = models.TextField(blank=True)
     heard_from = ChoiceArrayField(base_field=models.TextField(
             choices=string_choices([
@@ -212,11 +56,8 @@ class Demographic(models.Model):
     )
 
 
-class Quality(models.Model):
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    created = models.DateTimeField(auto_now_add=True)
-    modified = models.DateTimeField(auto_now=True)
-    event = models.ForeignKey("Event", on_delete=models.CASCADE)
+class Quality(EditTracking):
+    event = models.ForeignKey("Event", on_delete=models.CASCADE, related_name="quality")
     used_resources_before = models.TextField(
         blank=True,
         choices=[
@@ -282,7 +123,7 @@ class Quality(models.Model):
     )
 
 
-class Impact(models.Model):
+class Impact(EditTracking):
     HOW_LONG_CHOICES = [
         ("Less than 6 months", "Less than 6 months"),
         ("6 months to a year", "6 months to a year"),
@@ -377,10 +218,7 @@ class Impact(models.Model):
         ("Other", "Other"),
     ]
 
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    created = models.DateTimeField(auto_now_add=True)
-    modified = models.DateTimeField(auto_now=True)
-    event = models.ForeignKey("Event", on_delete=models.CASCADE)
+    event = models.ForeignKey("Event", on_delete=models.CASCADE, related_name="impact")
     when_attend_training = models.TextField(
         choices=HOW_LONG_CHOICES)
     main_attend_reason = models.TextField(blank=True, choices=REASON_CHOICES)
@@ -408,61 +246,3 @@ class Impact(models.Model):
 
     def __str__(self):
         return f"Attendance: {self.get_how_long_ago_display()}, Reason: {self.get_main_attend_reason_display()}, Use Before: {self.how_often_use_before}, Use After: {self.how_often_use_after}, Able to Explain: {self.able_to_explain}"
-
-
-class Node(models.Model):
-    name = models.TextField()
-    country = models.TextField()
-
-    def __str__(self):
-        return (
-            f"{self.name} ({self.country})"
-            if self.country
-            else self.name
-        )
-
-
-def is_ror_id(value):
-    if re.match("^https://ror.org/[a-zA-Z0-9]+$", value):
-        return value
-    raise ValidationError(f"Not a valid ror id: {value}")
-
-
-class OrganisingInstitution(models.Model):
-    name = models.TextField()
-    country = models.TextField()
-    ror_id = models.URLField(max_length=512, unique=True, null=True, validators=[is_ror_id])
-
-    def __str__(self):
-        return (
-            f"{self.name} ({self.country})"
-            if self.country
-            else str(self.name)
-        )
-
-    def get_absolute_url(self):
-        return reverse("institution-edit", kwargs={"pk": self.id})
-
-    def update_ror_data(self):
-        try:
-            ror_id_base = re.match("^https://ror.org/(.+)$", self.ror_id)[0]
-            ror_url = f"https://api.ror.org/organizations/{ror_id_base}"
-            response = requests.get(ror_url, allow_redirects=True)
-            if response.status_code == 200:
-                data = response.json()
-                self.name = data["name"]
-                self.country = data.get("country", {}).get("country_name", "")
-            else:
-                raise ValidationError(f"Could not fetch ROR data for: {self.ror_id}, {ror_url}, {response.status_code}")
-        except TypeError:
-            raise ValidationError(f"Not a valid ror id: {self.ror_id}")
-
-
-def get_node(self):
-    node_name = f"ELIXIR-{self.username.upper()}"
-    try:
-        return Node.objects.get(name=node_name)
-    except Node.DoesNotExist:
-        return None
-
-User.add_to_class("get_node", get_node)
